# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gym Multi-Tenant API",
        "description": "API documentation for the Laravel Gym Management system",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/branches": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get list of branches",
                "operationId": "ad820ebe136953fdc3366c5bc5cc54f1",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Optional filters",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Branches"
                ],
                "summary": "Create a new branch",
                "operationId": "e43f7bd6cc2b1cce23552a9fe71ba519",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBranchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/branches/{id}": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get branch by ID",
                "operationId": "016f17debb8d9d7b65ce115c6da4530e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Branches"
                ],
                "summary": "Update an existing branch",
                "operationId": "39931fc92262f0ef05caaf266aa631fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBranchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Branches"
                ],
                "summary": "Delete a branch",
                "operationId": "8fc1a339b6340b9bd2d108b1f3e16d68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cities": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get list of cities",
                "operationId": "19c198ca7a52b85bd48759713368b62c",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Optional filters",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cities"
                ],
                "summary": "Create a new city",
                "operationId": "65c199e29c92fb5c4461256a7eecebc5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "country_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "City created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cities/{id}": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get city by ID",
                "operationId": "c9e957f580378fa298b78507cf4a4c3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "City not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cities"
                ],
                "summary": "Update a city",
                "operationId": "41dedbb5c560aae5546829626129cd3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "country_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "City updated"
                    },
                    "404": {
                        "description": "City not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cities"
                ],
                "summary": "Delete a city",
                "operationId": "0f17db22f38516d8a8e0d5b9336e1770",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "City deleted"
                    },
                    "404": {
                        "description": "City not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreBranchRequest": {
                "required": [
                    "en_name",
                    "ar_name",
                    "gym_id",
                    "city_id",
                    "subscribe_date",
                    "expire_date",
                    "capacity",
                    "status"
                ],
                "properties": {
                    "en_name": {
                        "type": "string",
                        "example": "Main Branch"
                    },
                    "ar_name": {
                        "type": "string",
                        "example": "الفرع الرئيسي"
                    },
                    "gym_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "subscribe_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "expire_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 100
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "contact": {
                        "properties": {
                            "facebook": {
                                "type": "string",
                                "example": "https://facebook.com/branch"
                            },
                            "twitter": {
                                "type": "string",
                                "example": "https://twitter.com/branch"
                            }
                        },
                        "type": "object"
                    },
                    "addresses": {
                        "properties": {
                            "address": {
                                "type": "string",
                                "example": "123 Main St"
                            },
                            "latitude": {
                                "type": "number",
                                "format": "float",
                                "example": 30.0444
                            },
                            "longitude": {
                                "type": "number",
                                "format": "float",
                                "example": 31.2357
                            }
                        },
                        "type": "object"
                    },
                    "configs": {
                        "type": "object"
                    },
                    "commerces": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "UpdateBranchRequest": {
                "properties": {
                    "en_name": {
                        "type": "string",
                        "example": "Main Branch"
                    },
                    "ar_name": {
                        "type": "string",
                        "example": "الفرع الرئيسي"
                    },
                    "gym_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "subscribe_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "expire_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 100
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "contact": {
                        "properties": {
                            "facebook": {
                                "type": "string",
                                "example": "https://facebook.com/branch"
                            },
                            "twitter": {
                                "type": "string",
                                "example": "https://twitter.com/branch"
                            }
                        },
                        "type": "object"
                    },
                    "addresses": {
                        "properties": {
                            "address": {
                                "type": "string",
                                "example": "123 Main St"
                            },
                            "latitude": {
                                "type": "number",
                                "format": "float",
                                "example": 30.0444
                            },
                            "longitude": {
                                "type": "number",
                                "format": "float",
                                "example": 31.2357
                            }
                        },
                        "type": "object"
                    },
                    "configs": {
                        "type": "object"
                    },
                    "commerces": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}